# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    ChallengeResponse:
      type: object
      properties:
        nonce:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - nonce
        - expiresAt
    ChallengeRequest:
      type: object
      properties:
        stellarAddress:
          type: string
      required:
        - stellarAddress
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required:
        - accessToken
        - refreshToken
    VerifyRequest:
      type: object
      properties:
        stellarAddress:
          type: string
        nonce:
          type: string
        signature:
          type: string
      required:
        - stellarAddress
        - nonce
        - signature
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
        - refreshToken
    Market:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
        metadata:
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - status
        - version
        - createdAt
    MarketSearchResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Market'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        page:
          type: integer
        fallback:
          type: boolean
        pagination:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            page:
              type: integer
            total:
              type: integer
            fallback:
              type: boolean
          required:
            - limit
            - offset
            - page
            - total
            - fallback
        meta:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            page:
              type: integer
            total:
              type: integer
            fallback:
              type: boolean
          required:
            - limit
            - offset
            - page
            - total
            - fallback
      required:
        - data
        - total
        - limit
        - offset
        - page
        - fallback
        - pagination
        - meta
    PatchMarketRequest:
      type: object
      properties:
        question:
          type: string
        metadata:
          nullable: true
        expectedVersion:
          type: integer
          minimum: 0
      required:
        - expectedVersion
    PredictionCountResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            marketId:
              type: string
            count:
              type: integer
              minimum: 0
            computedAt:
              type: string
              format: date-time
            cached:
              type: boolean
          required:
            - marketId
            - count
            - computedAt
            - cached
      required:
        - data
...
    FeaturedMarket:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
        resolutionOutcome:
          type: string
          nullable: true
        resolutionTime:
          type: string
          format: date-time
        winningOutcome:
          type: string
          nullable: true
        metadata:
          nullable: true
        featuredAt:
          type: string
          nullable: true
          format: date-time
        featuredBy:
          type: string
          nullable: true
      required:
        - id
        - question
        - status
        - resolutionTime
        - featuredAt
        - featuredBy
...
paths:
...
  /api/auth/challenge:
    post:
      operationId: authChallenge
      tags:
        - Auth
      summary: Request a sign-in challenge nonce
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeRequest'
            examples:
              challengeRequest:
                value:
                  stellarAddress: GABC1234567890DEFGHIJKLMNOPQRSTUVWX
      responses:
        '201':
          description: Challenge issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
              examples:
                challengeIssued:
                  value:
                    nonce: challenge-nonce-001
                    expiresAt: '2026-07-25T12:00:00.000Z'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
  /api/auth/verify:
    post:
      operationId: authVerify
      tags:
        - Auth
      summary: Verify challenge signature and obtain JWT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
            examples:
              verifyRequest:
                value:
                  stellarAddress: GABC1234567890DEFGHIJKLMNOPQRSTUVWX
                  nonce: challenge-nonce-001
                  signature: ed25519-signature-hex
      responses:
        '200':
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
              examples:
                tokensIssued:
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJnb29nbGUtdXNlcjEifQ.signature
                    refreshToken: refresh-token-001
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '401':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/auth/refresh:
    post:
      operationId: authRefresh
      tags:
        - Auth
      summary: Rotate a refresh token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            examples:
              refreshTokenRequest:
                value:
                  refreshToken: refresh-token-001
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
              examples:
                refreshedTokens:
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJnb29nbGUtdXNlcjEifQ.signature
                    refreshToken: refresh-token-002
        '400':
          description: Missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Reuse detected — family revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/auth/logout:
    post:
      operationId: authLogout
      tags:
        - Auth
      summary: Revoke the entire refresh-token family
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            examples:
              logoutRequest:
                value:
                  refreshToken: refresh-token-001
      responses:
        '204':
          description: Logged out
        '400':
          description: Missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/markets/recommendations:
    get:
      operationId: getMarketRecommendations
      tags:
        - Markets
      summary: Get personalized market recommendations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of recommended markets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
                required:
                  - data
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/markets/search:
    get:
      operationId: searchMarkets
      tags:
        - Markets
      summary: Full-text search across markets
      parameters:
        - schema:
            type: string
            minLength: 1
          required: true
          name: q
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            default: 20
          required: false
          name: limit
          in: query
        - schema:
            type: integer
            nullable: true
            minimum: 0
            default: 0
          required: false
          name: offset
          in: query
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
          required: false
          name: page
          in: query
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketSearchResult'
        '400':
          description: Missing query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/markets/{id}:
    get:
      operationId: getMarketById
      tags:
        - Markets
      summary: Get a market by ID
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
                required:
                  - data
              examples:
                default:
                  value:
                    data:
                      id: market-001
                      question: Will the US win the 2026 FIFA World Cup?
                      status: active
                      metadata:
                        category: sports
                        resolutionSource: official
                      version: 1
                      createdAt: '2026-01-10T12:00:00.000Z'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    patch:
      operationId: updateMarket
      tags:
        - Markets
      summary: Update a market (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMarketRequest'
      responses:
        '200':
          description: Updated market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
                required:
                  - data
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/markets/{id}/prediction-count:
    get:
      operationId: getMarketPredictionCount
      tags:
        - Markets
      summary: Get total prediction count for a market
      description: >-
        Returns the total number of predictions placed on the given market. Results are cached in Redis for 60 seconds.
        The `cached` field indicates whether the value came from the cache.
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Prediction count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionCountResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/markets/featured:
    get:
      operationId: getFeaturedMarkets
      tags:
        - Markets
      summary: List currently featured markets for the home page
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 20
          required: false
          name: limit
          in: query
      responses:
        '200':
          description: Featured markets ordered by most recently featured first
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeaturedMarket'
                required:
                  - data
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...